feat(user_pool): update UserPoolStats to use u64 for connection metrics and add new statistics

+++ db/src/tiberius/user_pool.rs
@@ -10,12 +10,19 @@ use ::tracing::debug;
 /// Aggregated connection statistics across all active user pools.
 #[derive(Default)]
 pub struct UserPoolStats {
-    pub total: u32,
-    pub idle: u32,
-    pub used: u32,
+    /// Total number of active connections across all user pools.
+    pub total: u64,
+    /// Total number of idle connections across all user pools.
+    pub idle: u64,
+    /// Total number of used connections across all user pools.
+    pub used: u64,
+    /// Total number of connections created across all user pools.
     pub created: u64,
+    /// Total number of connections closed due to idle timeout across all user pools.
     pub closed_idle_timeout: u64,
+    /// Total number of connections closed due to reaching max lifetime across all user pools.
     pub closed_max_lifetime: u64,
+    /// Total number of connections closed due to errors across all user pools.
     pub closed_error: u64,
 }
 
@@ -138,9 +145,9 @@ impl UserPoolManager {
         let mut stats = UserPoolStats::default();
         for (pool, _) in guard.values() {
             let state = pool.state();
-            stats.total += state.connections;
-            stats.idle += state.idle_connections;
-            stats.used += state.connections.saturating_sub(state.idle_connections);
+            stats.total += state.connections as u64;
+            stats.idle += state.idle_connections as u64;
+            stats.used += state.connections.saturating_sub(state.idle_connections) as u64;
             stats.created += state.statistics.connections_created;
             stats.closed_idle_timeout += state.statistics.connections_closed_idle_timeout;
             stats.closed_max_lifetime += state.statistics.connections_closed_max_lifetime;
