+++ infoportal/src/http/monitoring.rs
@@ -46,9 +46,9 @@ impl Monitoring {
         Monitoring {
             db: Db {
                 connections: Connections {
-                    total: pool_state.connections + up.total,
-                    idle: pool_state.idle_connections + up.idle,
-                    used: (pool_state.connections - pool_state.idle_connections) + up.used,
+                    total: pool_state.connections as u64 + up.total,
+                    idle: pool_state.idle_connections as u64 + up.idle,
+                    used: (pool_state.connections - pool_state.idle_connections) as u64 + up.used,
                     created: pool_stats.connections_created + up.created,
                     closed_idle_timeout: pool_stats.connections_closed_idle_timeout + up.closed_idle_timeout,
                     closed_max_lifetime: pool_stats.connections_closed_max_lifetime + up.closed_max_lifetime,
@@ -268,11 +268,11 @@ impl From<&CacheStats> for Caches {
 #[serde(rename_all = "camelCase")]
 struct Connections {
     /// The total number of connections.
-    total: u32,
+    total: u64,
     /// The number of connections that are currently not in use.
-    idle: u32,
+    idle: u64,
     /// The number of connections that are currently actively being used.
-    used: u32,
+    used: u64,
     /// The number of connections that have been created.
     created: u64,
     /// The number of connections that have been closed due to a timeout.
